Render all CMS-pages from slug-endpoint
