+++ apps/store/src/services/storyblok/Storyblok.helpers.ts
@@ -1,7 +1,7 @@
 import { StoryblokClient } from '@storyblok/js'
 import { SbBlokData, StoryData } from '@storyblok/react'
 import { Language } from '@/utils/l10n/types'
-import { LinkField, StoryblokVersion } from './storyblok'
+import { LinkField, ProductStory, StoryblokVersion } from './storyblok'
 
 export const filterByBlockType = <BlockData extends SbBlokData>(
   blocks: BlockData[] = [],
@@ -61,4 +61,8 @@ const makeAbsolute = (url: string) => {
     return url
   }
   return '/' + url
+}
+
+export const isProductStory = (story: StoryData): story is ProductStory => {
+  return story.content.component === 'product'
 }
